requirements.txt
google-cloud-bigquery
google-cloud-pubsub
requests
google-cloud-logging